remove TestV2FileContractRenewedToFrom, remove old v2consensus_test.go and move helpers that we still use to the refactored file
